requirements.txt
accelerate>=0.30.0
black
cloudpickle
diffusers>=0.26.3
future
fvcore
huggingface_hub==0.23.4
hydra-core
iopath
matplotlib<=3.9.1
numpy>=1.24.4
omegaconf
opencv_python==4.10.0.84
pillow>=10.3.0
pycocotools
pydot
PyYAML==6.0.1
scipy<=1.13.1
setuptools
scikit-image<=0.24.0
tabulate
tensorboard
termcolor
tqdm<=4.66.4
transformers>=4.27.3
xformers>=0.0.23.post1
yacs
Ninja==1.11.1.1
git+https://github.com/facebookresearch/detectron2.git@v0.6
git+https://github.com/facebookresearch/detectron2.git@v0.6#subdirectory=projects/DensePose
 